+++ src/tests/test_definitions.py
@@ -5,6 +5,9 @@ import re
 import pytest
 from pydocstyle.violations import Error, ErrorRegistry
 from pydocstyle.checker import check
+from pydocstyle.config import ConfigurationParser
+
+DEFAULT_PROPERTY_DECORATORS = ConfigurationParser.DEFAULT_PROPERTY_DECORATORS
 
 
 @pytest.mark.parametrize('test_case', [
@@ -35,10 +38,14 @@ def test_complex_file(test_case):
     test_case_file = os.path.join(test_case_dir,
                                   'test_cases',
                                   test_case + '.py')
-    results = list(check([test_case_file],
-                         select=set(ErrorRegistry.get_error_codes()),
-                         ignore_decorators=re.compile(
-                             'wraps|ignored_decorator')))
+    results = list(
+        check(
+            [test_case_file],
+            select=set(ErrorRegistry.get_error_codes()),
+            ignore_decorators=re.compile('wraps|ignored_decorator'),
+            property_decorators=DEFAULT_PROPERTY_DECORATORS,
+        )
+    )
     for error in results:
         assert isinstance(error, Error)
     results = {(e.definition.name, e.message) for e in results}