+++ docs/release_notes.rst
@@ -8,6 +8,9 @@ Release Notes
 Current Development Version
 ---------------------------
 
+New Features
+
+* Add support for `property_decorators` config to ignore  D401
 
 6.1.1 - May 17th, 2021
 ---------------------------